+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,7 @@
+# Issue
+
+# Solution
+
+# Long explanation
+
+# Test
