+++ main.go
@@ -15,67 +15,6 @@ const (
 	CATALOG_ITEMS = "data-catalog-items-categories-request.json"
 )
 
-// Request Catalog
-type DataRequest struct {
-	Type       string     `json:"type"`
-	Attributes Attributes `json:"attributes"`
-}
-
-type CatalogRequest struct {
-	Data DataRequest `json:"data"`
-}
-
-// Request CatalogItem
-type CatalogItemAttributesRequest struct {
-	Name        string   `json:"name"`
-	Description string   `json:"description"`
-	Type        string   `json:"type"`
-	Slug        string   `json:"slug"`
-	Scope       []string `json:"scope"`
-	Parent      string   `json:"parent"`
-}
-
-type CatalogItemDataRequest struct {
-	Type       string                       `json:"type"`
-	Attributes CatalogItemAttributesRequest `json:"attributes"`
-}
-
-type CatalogItemRequest struct {
-	Data CatalogItemDataRequest `json:"data"`
-}
-
-// Response Catalogs Data
-type Attributes struct {
-	Name        string `json:"name"`
-	Description string `json:"description"`
-	Type        string `json:"type"`
-	CreatedAt   string
-	UpdatedAt   string
-}
-
-type Catalog struct {
-	Type       string
-	Id         string
-	Attributes Attributes
-}
-
-// Response Catalogs Metadata
-type Paginate struct {
-	PageCount int
-	Page      int
-	Limit     int
-}
-
-type Metadata struct {
-	Paginate Paginate
-}
-
-// Response Catalogs
-type CatalogsResponseGet struct {
-	Metadata Metadata
-	Data     []Catalog
-}
-
 // Control flow variables
 var (
 	token                 *string
@@ -136,89 +75,6 @@ func makeAPICalls(item CatalogItemRequest) {
 	response, err := ioutil.ReadAll(res.Body)
 
 	fmt.Printf("%s", response)
-}
-
-func createCatalogItems() {
-	jsonFile, err := os.Open(CATALOG_ITEMS)
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer jsonFile.Close()
-
-	jsonBytes, err := ioutil.ReadAll(jsonFile)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	json.Unmarshal(jsonBytes, &catalogItems)
-
-	for k, _ := range catalogItems {
-		catalogItems[k].Data.Attributes.Parent = catalogId
-	}
-}
-
-func getCatalogs() {
-	client := &http.Client{}
-
-	req, err := http.NewRequest(http.MethodGet, urlCatalogs, nil)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	req.Header.Add("Authorization", *token)
-
-	res, err := client.Do(req)
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer res.Body.Close()
-
-	body, err := ioutil.ReadAll(res.Body)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	json.Unmarshal(body, &catalogs)
-}
-
-func createCategoryCatalog() {
-	client := &http.Client{}
-
-	catalogRequest := CatalogRequest{
-		Data: DataRequest{
-			Type: "catalogs",
-			Attributes: Attributes{
-				Name:        "categories",
-				Description: "Categories of the CMS",
-				Type:        "categories",
-			},
-		},
-	}
-
-	data, err := json.Marshal(catalogRequest)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	req, err := http.NewRequest(http.MethodPost, urlCatalogs, bytes.NewBuffer(data))
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	req.Header.Add("Authorization", *token)
-
-	res, err := client.Do(req)
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer res.Body.Close()
-
-	body, err := ioutil.ReadAll(res.Body)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	json.Unmarshal(body, &catalog)
 }
 
 func getCategoryIdCatalog() {
